+++ client/data/constants.js
@@ -1,7 +1,7 @@
 // event types
 
 // IDs
-export const GET_ALL_ID = 'get-all';
+
 
 // classes
