+++ client/styles/index.css
@@ -1,20 +1,38 @@
 @import url('./other-file.css');
-.wrapper {
-    height: 800px;
-    text-align: center;
+.page {
     background-color: aqua;
 }
-.container h1{
-  padding: 40px;
-}
 
+.input {
+  display: block;
+  text-align: center;
+  height: 30%;
+  
+}
 .buttons
     {
     margin:100px;
-    display: flex;
     text-align: center;
-    justify-content: center;
+    padding: 0;
 }
+.buttons *{
+  display: inline-block;
+}
+
+#result {
+  display: grid;
+  grid-template: repeat(2, 300px) /repeat(5, 1fr);
+  grid-gap: 3px;}
+
 .user{
-  background-color: red;
+  background-color: lightblue;
+}
+.user h2{
+  padding-left: 25px;
+  text-align: left;
+}
+.user p{
+text-align: left;
+align-items: center;
+ padding-left: 25px;
 }
